+++ CHANGELOG.md
@@ -50,6 +50,7 @@ New devices:
 - Zemismart ZM-L03E-Z 3 gang with neutral wire switch
 - Zemismart RGB LED downlight
 - Zemismart Curtain/roller blind switch
+- Bitron Wireless wall thermostat with relay
 
 ## 0.0.20
 - Properly handle heartbeat message from IKEA TRADFRI Remote Control